requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.26.0
requests>=2.31.0
plotly>=5.20.0
python-dotenv>=1.0.0
streamlit-autorefresh>=0.0.1
cryptography>=42.0.0
